Making Math standalone
The math module is now standalone with no external dependencies.
